Integrate Home VIP Components with View + Router
 - Integrate HomeViewController with HomeInteractor.
 - Create AppNavigator to display first screen.
 - Create Navigator to manage navigation between views.
 - Create DropDown List.
 - Create ImageStore to cash images.

+++ RecipeSearch/Core/Models/Recipe.swift
@@ -19,7 +19,7 @@ struct RecipeModel : Codable
         imageURL,
         publisherURL: String?
     
-    let socialRank: Int?
+    let socialRank: Double?
     
     enum CodingKeys: String, CodingKey
     {
@@ -62,5 +62,5 @@ struct Recipe
         imageURL,
         publisherURL: String
     
-    let socialRank: Int
+    let socialRank: Double
 }
